Got unit tests working
Unit tests are now working, though there are insufficient tests.

Also renamed some things and created helper scripts for running
the tool and running the tests.

Discovered how to use PYTHONPATH to aid in testing from the
root directory of the project. BUT I assume PYTHONPATH is empty
or non-extant and that is a bad assumption. I need to restore it
if already has a value.