Added --just-raw, --ignore-transclusions options
Added option for post processing the output from a Markdown processor,
looking for raw file include statements and inserting the raw text.

Added option to ignore Markdown transclusion specifications (so that
mdmerge does the same thing that Marked 2 does); this allows the
MultiMarkdown processor to handle the transclusions.

updated the docs to match the new behavior.

Fixed a bug handling unicode input and output for a Python 2
environment.