Ignore MultiMarkdown metadata in included files
This addresses issue #3.

The metadata in the leading file is maintained in the output, but
the metadata in all subsequent files is eaten, not sent to the
output stream. This is true for a single input file (in which case
the metadata in the nested files would be ignored); for index files
and multiple input files from the CLI, a similar rule applies.

For an index file, or for multiple input files from CLI, only keep
metadata from the first file.

In the CLI case, I found an additiona bug; I had to insert a blank
line between the files, and had to pass in the
discardMetadata argument as True for all but first file.

Multiple tests where added for various metadata scenarios. I also
enabled a number of disabled CLI execute tests.